+++ nav-app/src/app/utilities/globals.ts
@@ -1,0 +1,5 @@
+'use strict';
+import appPackageInfo from "../../../package.json";
+
+export const nav_version: string = appPackageInfo.version;
+export const layer_version: string = "4.4";
