+++ nav-app/src/app/utilities/navigator.d.ts
